text2img/requirements.txt
torch>=2.0.0
diffusers>=0.25.0
transformers>=4.36.0
pymongo>=4.6.0
google-cloud-storage>=2.13.0
requests>=2.31.0
Pillow>=10.0.0
tqdm>=4.66.0  # Required by diffusers